Add files via upload
This project focuses on predicting prices using a dataset meticulously chosen for its relevance and uniqueness, ensuring it has not been previously utilized in class projects like CKD or Telco Churn. The dataset, sourced from a comprehensive and credible database, encompasses a variety of input variables directly influencing the price, the output variable of interest. A thorough descriptive analysis was conducted, revealing the dataset's dimensions including the number of rows and columns, the percentage of missing values per column, and key statistical measures such as means, medians, modes, variances, and correlations. Additionally, an in-depth exploration of outliers through visualizations like scatter plots and correlation plots provided insightful observations.

The preprocessing phase was critical for enhancing model accuracy and involved careful variable treatment. This included the elimination of redundant columns and rows, imputation strategies for handling missing values, and necessary transformations to ensure data quality. A meticulous variable selection process was also undertaken to identify the most predictive features. Multiple models were then fitted with varying parameters to capture the underlying patterns effectively. Each model's performance was evaluated using a set of indicators, and the best-performing model was selected based on its performance on a test dataset. This comprehensive approach not only ensures the reliability of the price predictions but also demonstrates a structured methodology for tackling predictive modeling problems.